4.23

+++ part4/bloglist/tests/blog_api.test.js
@@ -2,6 +2,7 @@ const mongoose = require('mongoose');
 const supertest = require('supertest');
 const app = require('../app');
 const Blog = require('../model/blog');
+const User = require('../model/user');
 
 const dummyData = [
   {
@@ -24,12 +25,23 @@ const dummyData = [
   }
 ];
 
-beforeEach(async () => {
+beforeAll(async () => {
   await Blog.deleteMany({});
   dummyData.forEach(async blogData => {
     let newBlog = new Blog(blogData);
     await newBlog.save();
   });
+
+  await User.deleteMany({});
+  await api
+    .post('/api/users')
+    .send({
+      name: 'User 1',
+      username: 'user1',
+      password: 'User@1'
+    })
+    .expect(201)
+    .expect('Content-Type', /application\/json/);
 });
 
 const api = supertest(app);
@@ -51,18 +63,41 @@ test('a valid blog can be added', async () => {
     'likes': 95
   };
 
+  const loginRes = await api.post('/api/login')
+    .send({
+      name: 'User 1',
+      username: 'user1',
+      password: 'User@1'
+    })
+    .expect(200)
+    .expect('Content-Type', /application\/json/);
+
+  expect(loginRes.body.token).toBeDefined();
+
   await api
     .post('/api/blogs')
+    .set('authorization', 'bearer ' + loginRes.body.token)
     .send(newBlog)
     .expect(201)
     .expect('Content-Type', /application\/json/);
 
   const response = await api.get('/api/blogs');
-
-  const contents = response.body.map(r => r.title);
-
   expect(response.body).toHaveLength(dummyData.length + 1);
-  expect(contents).toContain('Heretics Of Dune');
+});
+
+test('authorized blog adding', async () => {
+  const newBlog = {
+    'title': 'Heretics Of Dune',
+    'author': 'Frank Herbert',
+    'url': 'https://www.adlibris.com/fi/kirja/heretics-of-dune-9781473233799',
+    'likes': 95
+  };
+
+  await api
+    .post('/api/blogs')
+    .send(newBlog)
+    .expect(401)
+    .expect('Content-Type', /application\/json/);
 });
 
 test('blogs have ID attribute', async () => {
@@ -79,11 +114,23 @@ test('blog missing like default to 0', async () => {
     'url': 'https://www.adlibris.com/fi/kirja/children-of-dune-9781473233782',
   };
 
-  const returnedBlog = await api.post('/api/blogs').send(newBlog).expect(201).expect('Content-Type', /application\/json/);
+  const loginRes = await api.post('/api/login')
+    .send({
+      name: 'User 1',
+      username: 'user1',
+      password: 'User@1'
+    })
+    .expect(200)
+    .expect('Content-Type', /application\/json/);
 
-  const response = await api.get(`/api/blogs/${returnedBlog.body.id}`);
+  expect(loginRes.body.token).toBeDefined();
 
-  expect(response.body.likes).toBe(0);
+  const returnedBlog = await api.post('/api/blogs')
+    .set('authorization', 'bearer ' + loginRes.body.token)
+    .send(newBlog).expect(201)
+    .expect('Content-Type', /application\/json/);
+
+  expect(returnedBlog.body.likes).toBe(0);
 });
 
 test('blog missing title and url', async () => {
@@ -91,23 +138,49 @@ test('blog missing title and url', async () => {
     'author': 'Frank Herbert',
   };
 
-  await api.post('/api/blogs').send(newBlog).expect(400).expect('Content-Type', /application\/json/);
+  const loginRes = await api.post('/api/login')
+    .send({
+      name: 'User 1',
+      username: 'user1',
+      password: 'User@1'
+    })
+    .expect(200)
+    .expect('Content-Type', /application\/json/);
+
+  expect(loginRes.body.token).toBeDefined();
+
+  await api.post('/api/blogs')
+    .send(newBlog)
+    .set('authorization', 'bearer ' + loginRes.body.token)
+    .expect(400)
+    .expect('Content-Type', /application\/json/);
 });
 
 test('blog delete one', async () => {
-  const res = await api.get('/api/blogs');
-  const id = res.body[0].id;
-  await api.delete(`/api/blogs/${id}`);
+  jest.setTimeout(5000);
+  const loginRes = await api.post('/api/login')
+    .send({
+      name: 'User 1',
+      username: 'user1',
+      password: 'User@1'
+    })
+    .expect(200)
+    .expect('Content-Type', /application\/json/);
 
-  const resAfterDelete = await api.get('/api/blogs');
-  const result = resAfterDelete.body.filter(blog => blog.id === id);
-  expect(result.length).toBe(0);
-});
+  expect(loginRes.body.token).toBeDefined();
+
+  const blog = await Blog.find({users: loginRes.body.id});
+
+  const id = blog[0].id;
+  
+  const res = await api.delete(`/api/blogs/${id}`).set('authorization', 'bearer ' + loginRes.body.token).expect(200);
+  expect(res.body.message).toBe('deleted');
+}, 10000000);
 
 test('blog update one', async () => {
   const res = await api.get('/api/blogs');
   const id = res.body[0].id;
-  
+
   const updatedData = {
     'likes': 1000
   };
